app/requirements.txt
dissect>=3.22
fastapi>=0.135.2
opensearch-py>=3.1.0
orjson>=3.11.7
taskiq>=0.12.1
uvicorn>=0.42.0
taskiq-aio-pika>=0.6.0
taskiq-postgres[psycopg]>=0.7.0
sqlalchemy>=2.0.48
psycopg[python]>=3.3.3
